# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Luna AI API",
    "version": "1.0.0",
    "description": "API Documentation for Luna AI Streamer - AI-powered streamer with VTube Studio integration, emotion system, and Rock Paper Scissors game",
    "contact": {
      "name": "Luna AI Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local development server"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/luna/rps/betting/create": {
      "post": {
        "tags": ["RPS"],
        "summary": "Create betting room",
        "description": "Create a new betting room",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["wallet", "betAmount"],
                "properties": {
                  "wallet": {
                    "type": "string"
                  },
                  "betAmount": {
                    "type": "number",
                    "description": "Bet amount in Luna tokens"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Room created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "roomId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/luna/rps/betting/join": {
      "post": {
        "tags": ["RPS"],
        "summary": "Join betting room",
        "description": "Join an existing betting room",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["wallet", "roomId"],
                "properties": {
                  "wallet": {
                    "type": "string"
                  },
                  "roomId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Joined room",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/luna/rps/betting/submit": {
      "post": {
        "tags": ["RPS"],
        "summary": "Submit betting choice",
        "description": "Submit choice in betting room",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["wallet", "roomId", "choice"],
                "properties": {
                  "wallet": {
                    "type": "string"
                  },
                  "roomId": {
                    "type": "string"
                  },
                  "choice": {
                    "type": "string",
                    "enum": ["rock", "paper", "scissors"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Choice submitted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "result": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "AdminSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-secret",
        "description": "Admin secret key (set ADMIN_SECRET in .env)"
      },
      "PurchaseSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-purchase-secret",
        "description": "Purchase webhook secret (set PURCHASE_SECRET in .env)"
      }
    }
  }
}